backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
sentence-transformers==2.2.2
faiss-cpu==1.8.0
numpy==1.24.3
feedparser==6.0.10
torch==2.0.1
transformers==4.33.2
